Refact

+++ database/migrations/030_tb_jogadores_alter.ts
@@ -4,14 +4,22 @@ export default class Altertable extends BaseSchema {
   protected tableName = "tb_jogadores";
 
   public async up() {
-    this.schema.alterTable(this.tableName, (table) => {
-      table.integer("gold").notNullable().defaultTo(0);
-    });
+    const hasGold = await this.schema.hasColumn(this.tableName, "gold");
+
+    if (!hasGold) {
+      this.schema.alterTable(this.tableName, (table) => {
+        table.integer("gold").notNullable().defaultTo(0);
+      });
+    }
   }
 
   public async down() {
-    this.schema.alterTable(this.tableName, (table) => {
-      table.dropColumn("gold");
-    });
+    const hasGold = await this.schema.hasColumn(this.tableName, "gold");
+
+    if (hasGold) {
+      this.schema.alterTable(this.tableName, (table) => {
+        table.dropColumn("gold");
+      });
+    }
   }
 }
